#14 -> Refonte Header + Début responsive
-> Suppression du Header

-> Extraction du router du index.js

-> création du router.jsx

-> Création du Nav pour responsive mobile ( burger menu )

-> Refonte du style générale de chaque page

-> Fin du responsive à venir
