--- src/components/Header/Header.css
@@ -1,30 +1,0 @@
-.headerBloc {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: #F6F6F6;
-  position: fixed;
-  width: 100%;
-  top: 0;
-  height: 100px;
-  padding: 10px 0px;
-  box-shadow: 0px 0px 5px black;
-  z-index: 100;
-}
-
-.title {
-  display: flex;
-  align-items: center;
-  position: absolute;
-  margin-top: 10px;
-  left: 130px;
-}
-.title b {
-  font-size: 3.5rem;
-}
-
-.logo {
-  width: 100px;
-  height: 100px;
-  margin-left: 20px;
-}/*# sourceMappingURL=Header.css.map */
