+++ src/components/Nav/Nav.scss
@@ -1,46 +1,214 @@
-ul {
+.NavBar{
     display: flex;
-    justify-content: flex-end;
-    list-style-type: none;
-    padding-inline-start: 0;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: space-between;
+    position: fixed;
+    top: 0;
+    z-index: 99;
+    width: 100%;
+    min-height: 75px;
+    padding: 1rem;
+    background-color: #F6F6F6;
+    color: #F95959;
+    box-shadow: 0 2px 5px #252525;
+}
+
+.Logo {
+    display: flex;
+    align-items: center;
+    height: 75px;
+    width: 75px;
+
+    img {
+        width: 100%;
+        height: 100%;
+    }
+
+    b {
+        font-size: 2.5rem;
+        color: #252525;
+    }
+}
+
+.NavList {
+    display: flex;
+    list-style: none;
+    padding: 0;
     margin: 0;
-    margin-right: 20px;
 
-    li {
-        display: flex;
-        gap: 20px;
-        margin: 10px;
+    .link {
         font-size: 1.5rem;
-
-        a {
-            text-decoration: none;
-            text-shadow: 1px 1px 2px #b8b8b8;
+        color: #F95959;
+        text-decoration: none;
+        padding: 0 0.8rem;
+        text-shadow: 1px 1px 2px #b8b8b8;
 
             &:hover {
                 text-shadow: 2px 2px 4px #b8b8b8;
                 transform: scale(115%);
             }
+    }
+
+    .activated {
+        font-size: 1.5rem;
+        text-decoration: underline #F95959;
+        text-shadow: 1px 1px 2px #b8b8b8;
+        padding: 0 0.8rem;
+        color: #252525;
+    
+        &:hover {
+            text-shadow: 1px 1px 2px #b8b8b8;
+            transform: none;
         }
     }
 }
 
-.NavBloc {
-    display: flex;
-    align-items: center;
+.NavBurger {
+    display: none;
+    width: 75px;
+    height: 75px;
+    background: transparent;
+    border: none;
+    color: inherit;
 }
 
-.link {
-    text-decoration: none;
-    color: #F95959;
-}
+@media screen and(max-width:850px) {
+    .Logo {
+        height: 40px;
 
-.activated {
-    text-decoration: underline #F95959;
-    text-shadow: 1px 1px 2px #b8b8b8;
-    color: #252525;
+        b {
+            font-size: 1.5rem;
+        }
+    }
 
-    &:hover {
-        text-shadow: 1px 1px 2px #b8b8b8;
-        transform: none;
+    .NavList {
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        position: fixed;
+        width: 0;
+        right: -100vw;
+        bottom: 0;
+        height: 100vh;
+        padding: 2rem;
+        background-color: #F95959;
+        visibility: none;
+        transition: all 800ms ease-out;
+
+        .link {
+            display: block;
+            padding: 2rem;
+            font-size: 5vw;
+            color: #F6F6F6;
+            background-color: #F95959;
+        }
+
+        .activated {
+            display: block;
+            font-size: 5vw;
+            padding: 2rem;
+            text-decoration: none;
+            text-shadow: 1px 1px 2px #b8b8b8;
+            color: #252525;
+        
+            &:hover {
+                text-shadow: 1px 1px 2px #b8b8b8;
+                transform: none;
+            }
+        }
+    }
+
+    .NavBurger {
+        height: 40px;
+        width: 40px;
+        display: block;
+
+        &:hover {
+            cursor: pointer;
+        }
+
+        .BurgerLine, .BurgerLine::after, .BurgerLine::before {
+            display: block;
+            width: 40px;
+            height: 5px;
+            position: relative;
+            border-radius: 5px;
+            background-color: #F95959;
+            transition: all 500ms ease-in-out;
+        }
+
+        .BurgerLine::after, .BurgerLine::before {
+            content: "";
+            position: absolute;
+            left: 0;
+        }
+
+        .BurgerLine::before {
+            transform: translateY(-12px);
+        }
+
+        .BurgerLine::after {
+            transform: translateY(12px);
+        }
+    }
+
+    .showBar .NavList{        
+        right: 0;
+        width: 100vw;
+        visibility: visible;
+    }  
+    
+    .NavItem {
+        background: transparent;
+        transform: translateY(100vh);
+    }
+
+    .showBar .NavItem {
+        transform: translateY(0);
+    }
+
+    .showBar .Slide1 {
+        transition: all 1000ms ease-out;
+    }
+
+    .showBar .Slide2 {
+        transition: all 1200ms ease-out;
+    }
+
+    .showBar .Slide3 {
+        transition: all 1400ms ease-out;
+    }
+
+    .showBar .Slide4 {
+        transition: all 1600ms ease-out
+    }
+    
+    .NavItem::after {
+        content: "-";
+        display: block;
+        margin: 0 auto;
+        width: 2vw;
+        height: 4px;
+        background-color: #F6F6F6;
+    }
+
+    .NavItem:last-child::after {
+        display: none;
+    }
+
+    .showBar .BurgerLine {
+        background: transparent;
+        width: 0;
+    }
+
+    .showBar .BurgerLine::before {
+        transform: rotate(45deg);
+        background: #F6F6F6;
+    }
+
+    .showBar .BurgerLine::after {
+        transform: rotate(-45deg);
+        background: #F6F6F6;
     }
 }
