+++ src/pages/Home/Home.scss
@@ -1,3 +1,7 @@
+* {
+    box-sizing: border-box;
+}
+
 .HomePage {
     background-color: #F6F6F6;
 }
@@ -16,25 +20,25 @@
     padding-top: 200px;
 
     h1 {
-        font-size: 6rem;
+        font-size: 7vw;
         margin: 0;
         text-shadow: 2px 2px 4px #F6F6F6;
     }
 
     h2 {
         margin-top: 40px;
-        font-size: 3rem;
+        font-size: 3vw;
         text-shadow: 2px 2px 4px #F6F6F6;
     }
 
 }
 
 .descBloc {
-    width: 900px;
+    width: 80vw;
     text-align: center;
 
     p {   
-        font-size: 2rem;  
+        font-size: 2vw;  
         text-shadow: 2px 2px 4px #F6F6F6;
         margin-top: 60px;     
     }
@@ -119,21 +123,20 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    margin: 150px auto;
+    margin: 100px auto;
     gap: 50px;
 }
 
 .profilePicBloc {
     display: flex;
     border-radius: 25px;
-    width: 400px;
-    height: 400px;
+    width: 30vw;
+    height: 30vw;
     overflow: hidden;
     box-shadow: 2px 2px 5px #F95959;
 
     img {
         width: 100%;
-        height: auto;
     }
 }
 
@@ -141,10 +144,10 @@
     display: flex;
     flex-direction: column;
     gap: 20px;
-    width: 800px;
-    height: 360px;
+    width: 40vw;
+    height: 30vw;
     justify-content: space-around;
-    font-size: 1.5rem;
+    font-size: 1.5vw;
     border: solid 2px black;
     border-radius: 25px;
     padding: 20px;
